phishing_detector/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.5
asgiref==3.9.1
attrs==25.3.0
banal==1.0.6
beautifulsoup4==4.12.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==46.0.1
dataset==1.6.2
datasets==4.1.1
dill==0.4.0
dj-rest-auth==7.0.1
Django==5.2.6
django-allauth==65.11.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.9.0
greenlet==3.2.4
gunicorn==23.0.0
huggingface-hub==0.35.0
idna==3.10
joblib==1.5.1
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.6.4
multiprocess==0.70.16
nltk==3.8.1
numpy==1.26.4
oauthlib==3.3.1
packaging==25.0
pandas==2.1.3
pillow==11.3.0
propcache==0.3.2
psycopg2-binary==2.9.9
pyarrow==21.0.0
pycparser==2.23
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.1
six==1.17.0
soupsieve==2.7
SQLAlchemy==1.4.54
sqlparse==0.5.3
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
xxhash==3.5.0
yarl==1.20.1
